Fixed bug for when no path is found

+++ letter.cpp
@@ -109,7 +109,7 @@ int main(int argc, char *argv[])
 		}
 	}
 
-	for (unsigned int i = 1; i < path->size() - 1; i++) {
+	for (int i = 1; i < (int)path->size() - 1; i++) {
 		if (path->at(i)) {
 			delete path->at(i); path->at(i) = NULL;
 		}
